feat: working on ui logic

+++ src/handling_requests.py
@@ -6,25 +6,30 @@ import os
 # -- Manga Search Engine
 
 base_url = "https://api.mangadex.org"
+manga_title = None#manga_title #input("Manga title: ")
 
 
-manga_title = input("Manga title: ")
-manga_title_response = requests.get(f"{base_url}/manga?title={manga_title}")
+def get_manga_title(manga_title):
+    try:
+        manga_title_response = requests.get(f"{base_url}/manga?title={manga_title}")
+        if manga_title_response.status_code == 200:
+                manga_data = manga_title_response.json().get("data")
+        
+                manga_id = manga_data[0]["id"]
+                print("Manga id",manga_id)
+        else:
+            manga_id = None
+        return manga_id
+    except Exception as e:
+        print("DEBUG: ",e)
+        print("Does this manga exist?")
 
-def get_manga_title(manga_title,manga_title_response):
-    if manga_title_response.status_code == 200:
-            manga_data = manga_title_response.json().get("data")
-    
-            manga_id = manga_data[0]["id"]
-            print("Manga id",manga_id)
-    else:
-        manga_id = None
-
-    return manga_id
-manga_id = get_manga_title(manga_title,manga_title_response)
-
-feed_url = f"{base_url}/manga/{manga_id}/feed?translatedLanguage[]=en&order[chapter]=asc"
-feed_response = requests.get(feed_url)
+#manga_id = get_manga_title()
+try:
+    feed_url = f"{base_url}/manga/{manga_id}/feed?translatedLanguage[]=en&order[chapter]=asc"
+    feed_response = requests.get(feed_url)
+except Exception as e:
+    print(e)
 def get_manga_chapters(feed_response,):
     if feed_response.status_code == 200:
         chapters = feed_response.json()["data"]
@@ -43,7 +48,8 @@ def get_manga_chapters(feed_response,):
 
 
     return chapter_id,chapter_number,chapter_1_id,chapter_1_num
-chapter_id,chapter_number,chapter_1_id,chapter_1_num = get_manga_chapters(feed_response)
+#chapter_id,chapter_number,chapter_1_id,chapter_1_num = get_manga_chapters(feed_response)
+
 
 def get_server_data():
     server_response = requests.get(f"{base_url}/at-home/server/{chapter_id}")
@@ -55,7 +61,7 @@ def get_server_data():
         pages = server_data["chapter"]["data"]
     
     return pages,host,chapter_hash
-pages,host,chapter_hash = get_server_data()
+#pages,host,chapter_hash = get_server_data()
 
 
 def downloading_chapters(pages,chapter_1_num):
@@ -63,20 +69,21 @@ def downloading_chapters(pages,chapter_1_num):
     for num in range(chapter_1_num):
         num += 1
         folder_path = f"Mangadex/{manga_title}/Chapter_{num}"
-        os.makedirs(folder_path,exist_ok=True)
+        if not os.path.exists(folder_path):
+            os.makedirs(folder_path,exist_ok=True)
     
-        for page in pages:
-            image_url = f"{host}/data/{chapter_hash}/{page}"
-            image_response = requests.get(image_url)
-            with open(f"{folder_path}/{page}","wb") as file:
-                file.write(image_response.content)
-        print(f"Downloaded {len(pages)} pages.")
+            for page in pages:
+                image_url = f"{host}/data/{chapter_hash}/{page}"
+                image_response = requests.get(image_url)
+                with open(f"{folder_path}/{page}","wb") as file:
+                    file.write(image_response.content)
+            print(f"Downloaded {len(pages)} pages.")
     else:
         print("Path already exists.")
     
     
 
-downloading_chapters(pages,chapter_1_num)
+#downloading_chapters(pages,chapter_1_num)
